+++ src/partials/elixir.js
@@ -1,23 +1,22 @@
 module.exports = [
     {
         "scope": [
-        ],
-        "settings": {"foreground": "#d4d4d4"},
-    },
-    {
-        "scope": [
             "comment.wildcard.elixir",
             "comment.unused.elixir",
         ],
         "settings": {
-            "foreground": "#808080",
+            "foreground": "#505050",
             "fontStyle": "italic",
         },
     },
     {
         "scope": [
-        "comment.documentation.heredoc.elixir",
+            "punctuation.section.function.elixir",
         ],
+        "settings": {"foreground": "#d4d4d4"},
+    },
+    {
+        "scope": ["comment.documentation.heredoc.elixir"],
         "settings": {"foreground": "#608b4e"},
     },
     {
@@ -30,9 +29,7 @@ module.exports = [
         "settings": {"foreground": "#9cdcfe"},
     },
     {
-        "scope": [
-            "punctuation.separator.method.elixir",
-        ],
+        "scope": ["punctuation.separator.method.elixir"],
         "settings": {"foreground": "#c586c0"},
     },
     {
@@ -46,19 +43,65 @@ module.exports = [
     },
     {
         "scope": [
-            // "comment.wildcard.elixir",
-            // "comment.unused.elixir",
             "source.elixir.embedded.source punctuation.section.embedded.elixir",
             "punctuation.section.embedded.elixir",
             "punctuation.definition.variable.elixir",
             "variable.other.anonymous.elixir",
+            "punctuation.definition.constant.elixir",
+        ],
+        "settings": {"foreground": "#569cd6"},
+    },
+    {
+        "scope": ["support.function.variable.quoted.single.elixir"],
+        "settings": {"foreground": "#d7ba7d"},
+    },
+    // =========================================================================
+    // Erlang
+    // =========================================================================
+    {
+        "scope": [
+            "punctuation.definition.placeholder.erlang",
+            "constant.other.placeholder.erlang",
+        ],
+        "settings": {"foreground": "#d7ba7d"},
+    },
+    {
+        "scope": ["punctuation.separator.parameters.erlang"],
+        "settings": {"foreground": "#d4d4d4"},
+    },
+    {
+        "scope": [
+            "punctuation.section.directive.begin.erlang",
+            "punctuation.section.directive.end.erlang",
+            "punctuation.separator.module-function.erlang",
+            "punctuation.separator.record-field.erlang",
+            "punctuation.separator.list.erlang",
+            "keyword.operator.symbolic.erlang",
+            "punctuation.separator.clause-head-body.erlang",
+        ],
+        "settings": {"foreground": "#c586c0"},
+    },
+    {
+        "scope": [
+            "punctuation.separator.clauses.erlang",
+            "punctuation.terminator.function.erlang",
+        ],
+        "settings": {"foreground": "#808080"},
+    },
+    {
+        "scope": [
+            "punctuation.separator.function-arity.erlang",
+            "variable.language.omitted.erlang",
+            // "punctuation.separator.clause-head-body.erlang",
+            // "keyword.operator.symbolic.erlang",
         ],
         "settings": {"foreground": "#569cd6"},
     },
     {
         "scope": [
-            "support.function.variable.quoted.single.elixir",
+            "variable.other.erlang",
+            "constant.other.symbol.unquoted.erlang",
         ],
-        "settings": {"foreground": "#d7ba7d"},
+        "settings": {"foreground": "#9cdcfe"},
     },
 ];