+++ src/partials/misc.js
@@ -7,6 +7,28 @@ module.exports = [
         "settings": {"foreground": "#505050"},
     },
     // =========================================================================
+    // Properties/INI
+    // =========================================================================
+    {
+        "scope": [
+            "comment.line.number-sign.ini",
+            "punctuation.definition.entity.ini",
+        ],
+        "settings": {"foreground": "#505050"},
+    },
+    {
+        "scope": ["keyword.other.definition.ini"],
+        "settings": {"foreground": "#9cdcfe"},
+    },
+    // {
+    //     "scope": ["entity.name.section.group-title.ini"],
+    //     "settings": {"foreground": "#569cd6"},
+    // },
+    {
+        "scope": ["punctuation.separator.key-value.ini"],
+        "settings": {"foreground": "#c586c0"},
+    },
+    // =========================================================================
     // YAML
     // =========================================================================
     {